requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
arcade==3.2.0
    # via musicplayer (pyproject.toml)
attrs==25.3.0
    # via pytiled-parser
audioread==3.0.1
    # via pyacoustid
certifi==2025.7.9
    # via requests
cffi==1.17.1
    # via pymunk
charset-normalizer==3.4.2
    # via requests
idna==3.10
    # via requests
iso3166==2.1.1
    # via musicplayer (pyproject.toml)
musicbrainzngs==0.7.1
    # via musicplayer (pyproject.toml)
mutagen==1.47.0
    # via musicplayer (pyproject.toml)
pillow==11.0.0
    # via arcade
pyacoustid==1.3.0
    # via musicplayer (pyproject.toml)
pycparser==2.22
    # via cffi
pydub==0.25.1
    # via musicplayer (pyproject.toml)
pyglet==2.1.6
    # via arcade
pymunk==6.9.0
    # via arcade
pypresence==4.3.0
    # via musicplayer (pyproject.toml)
pytiled-parser==2.2.9
    # via arcade
rapidfuzz==3.13.0
    # via thefuzz
requests==2.32.4
    # via pyacoustid
thefuzz==0.22.1
    # via musicplayer (pyproject.toml)
typing-extensions==4.14.1
    # via pytiled-parser
urllib3==2.5.0
    # via requests
watchdog==6.0.0
    # via musicplayer (pyproject.toml)
